benchmark/requirements.txt
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.16.1
python-dotenv==1.0.0
redis==5.0.1
rich==13.6.0
shellingham==1.5.3
typer==0.9.0
typing_extensions==4.8.0
redblue==0.1.0
